Refatorando Menu Principal e Inicial, ajuste da UTF-8
