+++ src/main/java/com/invest7/Main.java
@@ -1,7 +1,8 @@
 package com.invest7;
 
-import com.invest7.util.ConnectionFactory;
-import com.invest7.view.MenuInicial;
+import com.invest7.view.menus.MenuInicial;
+import java.io.PrintStream;
+import java.nio.charset.StandardCharsets;
 
 public class Main {
     public static final String RESET = "\033[0m";
@@ -16,9 +17,12 @@ public class Main {
 
     public static void main(String[] args) {
 
+        System.setOut(new PrintStream(System.out, true, StandardCharsets.UTF_8));
+        System.setErr(new PrintStream(System.err, true, StandardCharsets.UTF_8));
         MenuInicial menuInicial = new MenuInicial();
+
         System.out.println(CIANO + " ||=========================================================||" + RESET);
-        System.out.println(CIANO + " ***********Olá. Bem vindo ao Invest7: *********** " + RESET);
+        System.out.println(CIANO + " *********** Olá. Bem vindo ao Invest7: *********** " + RESET);
 
         menuInicial.exibirMenuInicial();
         System.out.println();
