+++ src/main/java/com/invest7/model/produtos/Fiis.java
@@ -1,4 +1,5 @@
-package com.invest7.model;
+package com.invest7.model.produtos;
+
 public class Fiis extends Produto {
     private double aporte;
     private int qtdCotas;
