+++ src/main/java/com/invest7/view/LoginView.java
@@ -1,6 +1,7 @@
 package com.invest7.view;
 
-import com.invest7.controller.UserController;
+import com.invest7.controller.user.UserController;
+
 import java.util.Scanner;
 
 public class LoginView {
