+++ src/main/java/com/invest7/view/menus/simulacao/MenuSimulacaoPerfil.java
@@ -1,4 +1,4 @@
-package com.invest7.view;
+package com.invest7.view.menus.simulacao;
 
 import java.util.Scanner;
