+++ src/main/java/com/invest7/view/menus/simulacao/perfil/MenuPerfilModerado.java
@@ -1,26 +1,29 @@
-package com.invest7.view;
+package com.invest7.view.menus.simulacao.perfil;
+
 import com.invest7.controller.CalculadoraRFprovisoria;
-import com.invest7.controller.CalculadoraVariavel;
+
 import java.util.Scanner;
 
-public class MenuPerfilConservador {
+public class MenuPerfilModerado {
     static Scanner sc = new Scanner(System.in);
     static CalculadoraRFprovisoria cf = new CalculadoraRFprovisoria();
 
+    public static void simulacaoPerfilModerado() {
 
-    public static void simulacaoPerfilConservador() {
+        String[] pacote = {"CRA","LCI", "CDB", "Tesouro Selic", "Tesouro IPCA+",  // RF - 50%
+                "VIVA3", "LREN3", //Ações - 20%
+                "XPLG11", "TGAR11", "MXRF11"}; //FIIs - 30%
 
-        String[] pacote = {"CRI", "CRA", "LCA", "LCI", "CDB", "Tesouro Selic", "Tesouro IPCA+",  // RF - 70%
-                "VIVA3 ", //Ações - 10%
-                "XPLG11", "TGAR11"}; //FIIs - 20%
-
-        System.out.println("=====Simulação de Perfil Conservador=====");
+        System.out.println("=====Simulação de Perfil Moderado=====");
         System.out.println("Digite o número de meses para a simulação: ");
         int meses = sc.nextInt();
         double impRenda = cf.calculoImpRenda(meses);
+        exibeRF(impRenda, meses);
+
+        double impRendaRV = 0.15;
     }
 
-    public static void exibeRF(double impRenda, int meses) {
+    public static void exibeRF (double impRenda, int meses) {
         System.out.println("++++Renda Fixa++++");
         System.out.println("Insira o valor inicial de investimento: R$ ");
         double valorInicial = sc.nextDouble();
@@ -30,13 +33,9 @@ public class MenuPerfilConservador {
 
         System.out.println("+++Juros Pré Fixados+++");
 
-        System.out.println("--CRI/CRA--");
+        System.out.println("--CRA--");
         double rBrutoCRA = cf.calculoRendBruto(valorInicial, aporteRF, 0.07, meses);
-        cf.imprIsento(rBrutoCRA, 0.07);
-
-        System.out.println("--LCA--");
-        double rBrutoLCA = cf.calculoRendBruto(valorInicial, aporteRF, 0.1478, meses);
-        cf.imprIsento(rBrutoLCA,0.1478 );
+        cf.imprIsento(rBrutoCRA,0.07 );
 
         System.out.println("--LCI--");
         double rBrutoLCI = cf.calculoRendBruto(valorInicial, aporteRF, 0.1298, meses);
@@ -55,15 +54,17 @@ public class MenuPerfilConservador {
 
         System.out.println("+++Juros Pós Fixados+++");
 
-        System.out.println("\n--CRI/CRA--");
+        System.out.println("\n--CRA--");
         double rBrutoCRApos = cf.calculoRendBruto(valorInicial, aporteRF, 0.06, meses);
-        cf.imprIsento(rBrutoCRApos, 0.06);
+        cf.imprIsento(rBrutoCRApos,0.06 );
+
+        System.out.println("--LCA--");
         double rBrutoLCApos = cf.calculoRendBruto(valorInicial, aporteRF, 0.12, meses);
-        cf.imprIsento(rBrutoLCApos, 0.12);
+        cf.imprIsento(rBrutoLCApos,0.12 );
 
         System.out.println("--LCI--");
         double rBrutoLCIpos = cf.calculoRendBruto(valorInicial, aporteRF, 0.102, meses);
-        cf.imprIsento(rBrutoLCIpos, 0.102);
+        cf.imprIsento(rBrutoLCIpos,0.102);
 
         System.out.println("\n--CDB Agibank--");
         double rBrutoCDBpos = cf.calculoRendBruto(valorInicial, aporteRF, 0.105, meses);
@@ -76,3 +77,4 @@ public class MenuPerfilConservador {
         cf.imprNIsento(rBrutoTesSelic, rLiqTesSelic, impRenda, 0.1375);
     }
 }
+
